Use hand-built findloc() for older compilers
Some compilers (GCC 8) do not support `findloc()`. For those cases, CMake
will detect it and set `FORTRAN_COMPILER_SUPPORTS_FINDLOC` to `OFF` and
then we can use it to conditionally build a `findloc.F90` with a
brute-force `findloc()`.

NOTE: This `findloc()` does *NOT* support all of Fortran's
functionality. Only the `findloc(array,value)` in *ONE* dimension.