Fixes #135 - incorrect use of pointers
Each of 3 distinct data structures were unintentionally sharing a 1D
array accross 33 Fortran pointers.  The fix was to directly allocate
each pointer and eliminate the fixed intermediary.

Because "33" was a magic number throughout the source, a new integer
parameter NUM_TAUA was introduced.


This component is not exercised by the usual tests and thus all that
has been verified is that the new version compiles.