Code to simplify reading of forcings.
A new module is introduduced:  ForcingSpecification.F90 and
a new procedure read_atmospheric() that uses the module.

There are some tradeoffs to be made with these interfaces.  The use of
a derived type reduces the total number of arguments for the
per-variable calls and allows them to fit nicely on one line.   But,
the derived type is not particularly natural, nor does it save much.
And the naming leaves much to be desired.    A proper fix may require
refactoring the existing MAPL_ReadForcing() interfaces.

The inconsistent formatting of file names (3 digits vs 2) as well as
data structures for multiple bins (2D array vs array of 1D pinters)
adds to the overall complexity here.   The former is probably
something that should be allowed (someday we may have a use case with
hundreds or even thousands of bins), but do we really need to support
both bin storage schemes?

If we do want to support the array-of-1d-arrays it would probably be
useful to elevate the type as a 1st class MAPL derived type.  The
ForcingSpecification could then directly support that type and avoid
the explicit loop in the client code.