removed commented lines
The lines that were deleted were for building the GCM so that an offline tool "dbg_cnlsm_offline.F90" can be used to debug CatchCN.
The offline tool requires the GCM to write output that serves as input to the offline tool.
This writing is turned on via a C preprocessor flag "DBG_CNLSM_INPUTS".
The commented line was incorrect because it used the "DBG_CATCH_INPUTS" directive associated with the similar debug tool for Catch.
The offline debug functionality can be integrated into cmake at a later time.  It may still build manually at this time, although this has not been verified.