Bug fix for optional input arg "conserve_ice10_tzero" (StieglitzSnow.F90)
An optional input argument ("conserve_ice10_tzero") was used in an if statement, but the optional input argument may not be provided by the calling routine. 
In the GCM, the optional argument is provided in both calls to StieglitzSnow_relayer().   In the LDAS, however, the optional argument is not always provided. 
The fix is to use the local variable ("conserve_ice10_tzero_tmp"), which always exists, in the if statement.