added stencil to compute total_concentration, wired into GFDL_1M primary class
implemented in such a way to avoid modifications to existing translate tests. driver tests expects total_concentration, so did not want to change driver call signiture.

technically, this could be done in setup, instead of the current location, but 1. setup does not currently expect any concentration, did not want to add new inputs and 2. ice/liquid concentrations are used in phase change - read only in the current state but no guarenteed in the future.

an independent stencil at the GFDL_1M level isn't necessarily the prettiest solution, but it is perhaps the most effective.