First attempt at implementing the downdraft_origin_level section - does not validate
Current implementation does not validate, but is probably short lived. This version is very brittle and should be reworked before the cumulus parameterization can be considered finished.

Reworking (with a proper implementation) probably requires the solver feature - or some other way of handling a while loop between I and K loops.

This implementation was started in case it the solver feature gets pushed back, in which case someone can come back here and clean up the current setup as a TEMPORARY implementation so that the convection scheme as a whole can progress.