First attempt at UpdraftMassFlux - does not validate.
lots of little nuanced errors and difficult to replicate patterns in here make it extremely difficult to fully implement, so only implemented the portion that runs in the current test case.

Implemented the mid/deep ZERO_DIFF=0 option and shallow (which does not run but is reasonable to assume will at some point). Shallow is currently disabled, lots of errors (most of which are good examples for errors that need improvement). Need to come back to this at some point